blog/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
cloudinary==1.44.1
Django==5.2.6
django-ckeditor==6.7.3
django-cloudinary-storage==0.3.0
django-js-asset==3.1.2
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0